requirements.txt
Flask==1.1.2
Flask-Cors==3.0.9
Flask-SocketIO==4.3.1
gevent-websocket==0.10.1
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
numba==0.51.2
numpy==1.18.5
opencv-contrib-python-headless
Pillow==8.0.1
tensorflow-cpu==2.3.1